+++ src/delay.rs
@@ -3,7 +3,6 @@
 //! [`embedded-hal`]: https://docs.rs/embedded-hal
 
 use cast::u64;
-use core::convert::Infallible;
 use embedded_hal::delay::DelayUs;
 use std::thread;
 use std::time::Duration;
@@ -12,13 +11,10 @@ use std::time::Duration;
 pub struct Delay;
 
 impl DelayUs for Delay {
-    type Error = Infallible;
-
-    fn delay_us(&mut self, n: u32) -> Result<(), Self::Error> {
+    fn delay_us(&mut self, n: u32) {
         let secs = n / 1_000_000;
         let nsecs = (n % 1_000_000) * 1_000;
 
         thread::sleep(Duration::new(u64(secs), nsecs));
-        Ok(())
     }
 }